We landed a flag to disable test utils in many builds but we need to fork the entrypoint to make it work with tests properly
